Level 1 and Level 2 Quiz Preparation/03 OpenAI Agents SDK_ModelSetting class/pyproject.toml
[project]
name = "02-openai-agents-sdk-modelsetting-class"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.1",
    "python-dotenv>=1.1.1",
]
